refactor(commands): extract /review command into review.rs
